+++ src/backends/x11.rs
@@ -22,7 +22,7 @@ use std::{
     fmt,
     fs::File,
     io, mem,
-    num::{NonZeroU16, NonZeroU32},
+    num::NonZeroU32,
     ptr::{null_mut, NonNull},
     slice,
     sync::Arc,
@@ -149,8 +149,11 @@ pub struct X11Impl<D: ?Sized, W: ?Sized> {
     /// Buffer has been presented.
     buffer_presented: bool,
 
-    /// The current buffer width/height.
-    size: Option<(NonZeroU16, NonZeroU16)>,
+    /// The current buffer width.
+    width: u16,
+
+    /// The current buffer height.
+    height: u16,
 
     /// Keep the window alive.
     window_handle: W,
@@ -300,7 +303,8 @@ impl<D: HasDisplayHandle + ?Sized, W: HasWindowHandle> SurfaceInterface<D, W> fo
             visual_id,
             buffer,
             buffer_presented: false,
-            size: None,
+            width: 0,
+            height: 0,
             window_handle: window_src,
         })
     }
@@ -310,7 +314,7 @@ impl<D: HasDisplayHandle + ?Sized, W: HasWindowHandle> SurfaceInterface<D, W> fo
         &self.window_handle
     }
 
-    fn resize(&mut self, width: NonZeroU32, height: NonZeroU32) -> Result<(), SoftBufferError> {
+    fn resize(&mut self, width: u32, height: u32) -> Result<(), SoftBufferError> {
         tracing::trace!(
             "resize: window={:X}, size={}x{}",
             self.window,
@@ -319,22 +323,18 @@ impl<D: HasDisplayHandle + ?Sized, W: HasWindowHandle> SurfaceInterface<D, W> fo
         );
 
         // Width and height should fit in u16.
-        let width: NonZeroU16 = width
-            .try_into()
-            .or(Err(SoftBufferError::SizeOutOfRange { width, height }))?;
-        let height: NonZeroU16 = height.try_into().or(Err(SoftBufferError::SizeOutOfRange {
-            width: width.into(),
-            height,
-        }))?;
+        let (width, height) = util::convert_size::<u16>(width, height)
+            .map_err(|_| SoftBufferError::SizeOutOfRange { width, height })?;
 
-        if self.size != Some((width, height)) {
+        if self.width != width && self.height != height {
             self.buffer_presented = false;
             self.buffer
-                .resize(self.display.connection(), width.get(), height.get())
+                .resize(self.display.connection(), width, height)
                 .swbuf_err("Failed to resize X11 buffer")?;
 
             // We successfully resized the buffer.
-            self.size = Some((width, height));
+            self.width = width;
+            self.height = height;
         }
 
         Ok(())
@@ -354,16 +354,13 @@ impl<D: HasDisplayHandle + ?Sized, W: HasWindowHandle> SurfaceInterface<D, W> fo
             depth: self.depth,
             buffer: &mut self.buffer,
             buffer_presented: &mut self.buffer_presented,
-            size: self.size,
+            width: self.width,
+            height: self.height,
         })
     }
 
     fn fetch(&mut self) -> Result<Vec<u32>, SoftBufferError> {
         tracing::trace!("fetch: window={:X}", self.window);
-
-        let (width, height) = self
-            .size
-            .expect("Must set size of surface before calling `fetch()`");
 
         // TODO: Is it worth it to do SHM here? Probably not.
         let reply = self
@@ -374,8 +371,8 @@ impl<D: HasDisplayHandle + ?Sized, W: HasWindowHandle> SurfaceInterface<D, W> fo
                 self.window,
                 0,
                 0,
-                width.get(),
-                height.get(),
+                self.width,
+                self.height,
                 u32::MAX,
             )
             .swbuf_err("Failed to send image fetching request")?
@@ -404,16 +401,17 @@ pub struct BufferImpl<'a> {
     depth: u8,
     buffer: &'a mut Buffer,
     buffer_presented: &'a mut bool,
-    size: Option<(NonZeroU16, NonZeroU16)>,
+    width: u16,
+    height: u16,
 }
 
 impl BufferInterface for BufferImpl<'_> {
-    fn width(&self) -> NonZeroU32 {
-        self.size.unwrap().0.into()
+    fn width(&self) -> u32 {
+        self.width as u32
     }
 
-    fn height(&self) -> NonZeroU32 {
-        self.size.unwrap().1.into()
+    fn height(&self) -> u32 {
+        self.height as u32
     }
 
     #[inline]
@@ -438,10 +436,6 @@ impl BufferInterface for BufferImpl<'_> {
 
     /// Push the buffer to the window.
     fn present_with_damage(self, damage: &[Rect]) -> Result<(), SoftBufferError> {
-        let (surface_width, surface_height) = self
-            .size
-            .expect("Must set size of surface before calling `present_with_damage()`");
-
         tracing::trace!("present: window={:X}", self.window);
 
         match self.buffer {
@@ -454,8 +448,8 @@ impl BufferInterface for BufferImpl<'_> {
                         xproto::ImageFormat::Z_PIXMAP,
                         self.window,
                         self.gc,
-                        surface_width.get(),
-                        surface_height.get(),
+                        self.width,
+                        self.height,
                         0,
                         0,
                         0,
@@ -481,8 +475,8 @@ impl BufferInterface for BufferImpl<'_> {
                                     u16::try_from(rect.y).ok()?,
                                     i16::try_from(rect.x).ok()?,
                                     i16::try_from(rect.y).ok()?,
-                                    u16::try_from(rect.width.get()).ok()?,
-                                    u16::try_from(rect.height.get()).ok()?,
+                                    u16::try_from(rect.width).ok()?,
+                                    u16::try_from(rect.height).ok()?,
                                 ))
                             })(
                             )
@@ -491,8 +485,8 @@ impl BufferInterface for BufferImpl<'_> {
                                 .shm_put_image(
                                     self.window,
                                     self.gc,
-                                    surface_width.get(),
-                                    surface_height.get(),
+                                    self.width,
+                                    self.height,
                                     src_x,
                                     src_y,
                                     width,
@@ -524,15 +518,13 @@ impl BufferInterface for BufferImpl<'_> {
     }
 
     fn present(self) -> Result<(), SoftBufferError> {
-        let (width, height) = self
-            .size
-            .expect("Must set size of surface before calling `present()`");
-        self.present_with_damage(&[Rect {
+        let rect = Rect {
             x: 0,
             y: 0,
-            width: width.into(),
-            height: height.into(),
-        }])
+            width: self.width.into(),
+            height: self.height.into(),
+        };
+        self.present_with_damage(&[rect])
     }
 }