+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added `Buffer::pixel_rows()` for iterating over rows of the buffer data.
 - Added `Buffer::pixels_iter()` for iterating over each pixel with its associated `x`/`y` coordinate.
 - **Breaking:** Removed generic type parameters `D` and `W` from `Buffer<'_>` struct.
+- **Breaking:** Changed `NonZeroU32` to `u32`, and handles zero-sized buffers internally.
 
 # 0.4.7